Move xcodeprojConfiguration check back to allPRs but enforce it only on macos-browser repository
