+++ org/allPRs.ts
@@ -1,4 +1,4 @@
-import {warn, danger} from "danger"
+import {fail, warn, danger} from "danger"
 
 export const prSize = async () => {  
     // Warn when there is a big PR
@@ -16,8 +16,29 @@ export const internalLink = async () => {
     }
 }
 
+export const xcodeprojConfiguration = async () => {
+    if (danger.github.pr.repo == "macos-browser") {
+        const projectFile = "DuckDuckGo.xcodeproj/project.pbxproj";
+        if (danger.git.modified_files.includes(projectFile)) {
+            let diff = await danger.git.diffForFile(projectFile);
+            let addedLines = diff.added.split(/\n/);
+            // The regex is equal to:
+            // * plus sign
+            // * 1 or more tabulation keys
+            // * an identifier (key) consisting of capital letters, underscores and digits,
+            // * a space and an equality sign
+            // * arbitrary number of any characters (the value can be empty)
+            // * a semicolon
+            if (addedLines.find(value => /^\+\t+[A-Z_0-9]* =.*;$/i.test(value))) {
+                fail("No configuration is allowed inside Xcode project file - use xcconfig files instead.");
+            }
+        }
+    }
+}
+
 // Default run
 export default async () => {
     await prSize()
     await internalLink()
+    await xcodeprojConfiguration()
 }
